cast int64 to smaller types (trunk only)
The devteam feedback was to place casts in the code
in question.

This puts explicit casts on some code that was being
compiled into 'int64' then stuffed into smaller types with
VC2005.

+++ src/artifact.c
@@ -212,7 +212,7 @@ register boolean mod;
 	if (otmp && *name)
 	    for (a = artilist+1; a->otyp; a++)
 		if (a->otyp == otmp->otyp && !strcmp(a->name, name)) {
-		    register int m = a - artilist;
+		    register int m = (int)(a - artilist);
 		    otmp->oartifact = (char)(mod ? m : 0);
 		    otmp->age = 0;
 		    if(otmp->otyp == RIN_INCREASE_DAMAGE)
